refactoring: deleted unused components and hooks from module/users.
